getNext8Bits()
Wrote a repeatable function to pull and return the next byte from the file. It's implementation into some 8 and 16 bit functions is now starting, and opcode 0x1, LD BC,d16, now has full and accurate use of this opcode. This includes inline printing of the hex argument to this opcode.