isAddressSpecial()
Over the past few days, I've been doing open-source research on the special addresses inside of a .gb file (which is just an image of a gameboy cartridge). Implemented function isAddressSpecial() to allow for checks. So far, this address will flag the entry point of the binary, but it'll also show the nintendo logo's location, as well as the game's title. These locations (except the OEP) aren't disassembled, as they're just data.